requirements.txt
fastapi~=0.103.2
fastapi-events~=0.10.1
mysql-connector-python~=8.1.0
pandas~=2.1.4
playwright~=1.40.0
python-dotenv~=1.0.0
logging~=0.4.9.6
SQLAlchemy~=2.0.25
uvicorn~=0.23.2
